Merge latest commit by the Admin

+++ doomclassic/doom/vars.h
@@ -1188,7 +1188,7 @@ std::vector <std::unique_ptr<skyflatmap_t>> skyFlatMaps;
 std::vector <std::unique_ptr<sky_t>> skies;
 int           customSkyIndex; //GK: ID24 SKYDEFS Skies
 //GK: Fire Sky Start
-int lastFireTime; //The Last time we run the fire simulation in milliseconds
+float lastFireTime; //The Last time we run the fire simulation in milliseconds
 bool initFire; //Flag to check if the fire sky is set up
 byte fireBuffer[FIRE_WIDTH * FIRE_HEIGHT]; //Buffer to store the fire calculations
 byte finalFireBuffer[FIRE_WIDTH * FIRE_HEIGHT]; //Rendering buffer for the fire Sky (only one column)
